requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
torch==2.1.0
torchvision==0.16.0
psycopg2-binary==2.9.9
pgvector==0.2.3
python-multipart==0.0.6
Pillow==10.0.1
numpy==1.26.0
python-dotenv==1.0.0
